requirements.txt
beautifulsoup4
igraph
joblib
leidenalg
matplotlib
numpy
pandas
pillow
psutil
scikit-learn
scipy
seaborn
shap
torch
torchvision
tqdm
transformers
umap-learn
vllm
xgboost
pyyaml